OSSM-755: Fix TLS certs used in tests for feature security.egress.tls
Integration tests for feature `security.egress.tls.*` were failing,
because of failing TLS handshake with an alert message "UNKNOWN CA (48)".
The reason of this failure is that certificates used in these tests
don't pass verification when OpenSSL 1.1.1k or higher is used.
Those certificates pass verification on Ubuntu 20.04 which delivers
OpenSSL 1.1.1f  31 Mar 2020, but do not pass on CentOS Stream 8
or RHEL 8.6 which deliver OpenSSL 1.1.1k  FIPS 25 Mar 2021.

I noticed that root-cert.pem and cert-chain.pem had specified
the same common name, so as a workaround I added SAN as a prefix
to CN in the cert-chain.pem.

Signed-off-by: Jacek Ewertowski <jewertow@redhat.com>